Move the domain modules into a domain package
