+++ toy_settings/repositories/memory.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from toy_settings import events
-from toy_settings import projections
-from toy_settings import storage
+from toy_settings.domain import events
+from toy_settings.domain import projections
+from toy_settings.domain import storage
 
 
 class MemoryRepo(storage.Repository):
